refactor: create and use the input module
