+++ packages/core/src/input/index.ts
@@ -1,0 +1,2 @@
+export * from './input-file-collection';
+export * from './input-file-resolver';
